==> notebooks/01 - Extract Addresses.ipynb <==
### Import

import os
import sys
import pandas as pd
import numpy as np

### Get Available Files

myfiles = []
for root, dirs, files in os.walk("../data/raw/", topdown=False):
    for name in files:
        if ('DS' not in name):
            myfiles.append(os.path.join(root, name))

### Last line has zipcode and state

def lastLineHasZipCodeAndState(lines):
    
    lastline = lines[-1]
    if (len(lastline) > 8):
        zipcode = lastline[-5:]
        if (zipcode.isdigit() == True and 'NY' in lastline):
            return True
        else:
            return False
    else:
        return False

### Clean lines

def cleanLines(lines):
    
    #Clean
    toParse = lines[:-3]
    toKeep = lines[-3:]
    returnlines = []
    for line in toParse:
        elements = line.split(' ')
        if ('.' in elements[0]):
            returnlines = [' '.join(elements[1:])]
    returnlines = returnlines + toKeep

    #Return
    return returnlines

### Parse Files to etract name, address and zipcode

for myfile in myfiles:
    bag = []
    lines = []
    print(myfile)
    with open(myfile,'r') as f:
        for i in range(8):
            line = f.readline()
            lines.append(line.rstrip())
        for line in f:
            lines = lines[1:]
            lines.append(line.rstrip())
            if (lastLineHasZipCodeAndState(lines) == True):
                newlines = cleanLines(lines)
                bag.append(newlines)
    savefile = myfile.replace('raw','extracted')
    with open(savefile,'w') as f:
        for item in bag:
            for line in item:
                f.write(line + '\n')
            f.write('\n')
