==> notebooks/02 - Chains to Dataframe.ipynb <==
### Import

import os
import sys
import pandas as pd
import numpy as np

### Get Available Files

myfiles = []
for root, dirs, files in os.walk("../data/processed/", topdown=False):
    for name in files:
        if ('DS' not in name):
            myfiles.append(os.path.join(root, name))

### Parse Files

bags = []
for myfile in myfiles:
    with open(myfile,'r') as f:
        bag = []
        for line in f:
            line = line.rstrip()
            if (len(line) > 0):
                bag.append(line)
            else:
                bags.append(bag)
                bag = []

### Create Dataframe

data = []
for bag in bags:
    chain = bag[0]
    phoneno = bag[1]
    address = bag[2]
    line4 = bag[3]
    elements = line4.split(',')
    city = elements[0]
    newelements = elements[1].split(' ')
    state = newelements[1]
    zipcode = newelements[2]
    data.append((chain,address,phoneno,city,state,zipcode))
data = pd.DataFrame(data)
data.columns = ['chain','address','phoneno','city','state','zipcode']

### Store

data.to_csv('../data/tables/chains.csv',index=False)
