==> notebooks/03 - Remove Chains outside NYC.ipynb <==
### Import

import os
import sys
import pandas as pd
import numpy as np

### Load Streets

zipcodes = pd.read_csv('../data/tables/nyc_zipcodes.csv')
zipcodes = zipcodes['ZCTA5CE10'].values.tolist()
zipcodes.append(10055)
zipcodes.append(10118)
zipcodes.append(10158)
zipcodes.append(10107)
zipcodes.append(10104)
zipcodes.append(10105)
zipcodes.append(10123)
zipcodes.append(10281)
zipcodes.append(11249)
zipcodes.append(10121)
zipcodes.append(10106)
zipcodes.append(11241)
zipcodes.append(11243)
zipcodes.append(11252)
zipcodes.append(11439)

### Load Chains

data = pd.read_csv('../data/tables/chains.csv')

### Determine which records to keep

mylist = []
indices = []
for i in range(data.shape[0]):
    dp = data.iloc[i]
    thiscode = dp.zipcode
    if thiscode in zipcodes:
        indices.append(i)

### Restrict and safe

data = data.iloc[indices]
data.to_csv('../data/tables/chains_cleaned.csv',index=False)
