==> notebooks/05 - Address to LatLng.ipynb <==
### Import

import os
import sys
import pandas as pd
import numpy as np
import re
import shapefile
from tqdm import tqdm

### Fix Record Helper Function

def fixRecord(val):
    
    #Fix
    elements = val.split('-')
    if (len(elements) > 1):
        val = int(str(int(elements[0])) + str(int(elements[1])))
    
    #Return
    return val

### Load Chain Data

data = pd.read_csv('../data/tables/chains_with_line_indices.csv')

### Load Street Data

sf = shapefile.Reader('../data/shapefiles/nyc_streets/nyc_streets_cleaned.shp')
shapeRecs = sf.shapeRecords()

### Get Lat/Lng Coordinates from Shapefile

lats = []
lngs = []
for i in tqdm(range(data.shape[0])):
    
    #Get Chain Information
    dp = data.iloc[i]
    index = dp.stindex
    
    #Find Shape + Record in Street Shapefile
    shaperec = shapeRecs[index]
    shape = shaperec.shape
    record = shaperec.record

    #If Straight line, interpolate
    if (len(shape.points) == 2):
        val = []
        val.append(int(fixRecord(record[3])))
        val.append(int(fixRecord(record[4])))
        val.append(int(fixRecord(record[7])))
        val.append(int(fixRecord(record[8])))
        val = np.array(val)
        if (dp.lat == 0 and record[5] != str(dp.zipcode) and record[9] != str(dp.zipcode)):
            print(dp.address,dp.zipcode,' -> ',np.min(val),np.max(val),record[12],record[5],record[9])
        indices = np.where(val > 0)[0]
        val = val[indices]
        maxval = float(np.max(val))
        minval = float(np.min(val))
        number = float(dp.number)
        number = number - minval
        maxval = maxval - minval
        minval = minval - minval
        if (maxval == 0): maxval = 1.0
        number = number / maxval
        minval = minval / maxval
        maxval = maxval / maxval
        ratio1 = 1 - number
        ratio2 = number
        spos = shape.points[0]
        epos = shape.points[1]
        slng = spos[0]
        slat = spos[1]
        elng = epos[0]
        elat = epos[1]
        lng = slng*ratio1 + elng*ratio2
        lat = slat*ratio1 + elat*ratio2

    if (len(shape.points) > 2):
        val = []
        val.append(int(fixRecord(record[3])))
        val.append(int(fixRecord(record[4])))
        val.append(int(fixRecord(record[7])))
        val.append(int(fixRecord(record[8])))
        val = np.array(val)
        points = np.array(shape.points)
        points = np.mean(points,axis=0)
        if (dp.lat == 0 and record[5] != str(dp.zipcode) and record[9] != str(dp.zipcode)):
            print(dp.address,dp.zipcode,' -> ',np.min(val),np.max(val),record[12],record[5],record[9])
        lng = points[0]
        lat = points[1]

    #If fixed lat/lng pair already provided
    if (dp.lat > 0):
        lng = dp.lng
        lat = dp.lat
    
    #Add
    lats.append(lat)
    lngs.append(lng)

#Store
data['lat'] = lats
data['lng'] = lngs   
data.to_csv('../data/tables/chains_with_lat_lng.csv',index=False)
