==> notebooks/11 - Graph Distances - Chains -> Tracts - Complex.ipynb <==
### Import

#%pip install pandas
#%pip install shapely
#%pip install networkx
#%pip install matplotlib
#%pip install scipy
#%pip install tqdm
#%pip install pyshp

### Import

import os
import sys
import pandas as pd
import numpy as np
import shapefile
import re
import pickle
import time
import networkx as nx
import matplotlib.pyplot as plt
from scipy.spatial import KDTree
from tqdm import tqdm
from collections import Counter
from scipy.spatial.distance import cdist
from scipy.spatial.distance import pdist

### Load Tract Data

tractdata = pd.read_csv('../data/acs/nyc_tracts/nyc_tracts_census_geo_node.csv')

### Load Chain Data

chaindata = pd.read_csv('../data/tables/chains_with_nodes.csv')

### Load Graph

X = pickle.load(open("../data/graphs/nyc.p","rb"))
O = X['O']
N = X['N']
edges = X['edges']
nodes = X['nodes']
nodenames = X['nodenames']
edgedatabase = X['edgedatabase']
streets = X['streets']
G = X['G']
Nlist = X['Nlist'].tolist()
pickle.dump(X,open('../data/graphs/nyc.p',"wb"))

### Create Tract Index List

tractlist = []
for i in range(tractdata.shape[0]):
    tractdp = tractdata.iloc[i]
    tract_node_name = tractdp.node_name.split(' | ')
    tract_node_name = ' | '.join(tract_node_name[1:])
    index = Nlist.index(tract_node_name)
    tractlist.append(index)
tractlist = np.array(tractlist)
tractdata['node_index'] = tractlist

### Create Chain Index List

chainlist = []
chainnames = []
for i in range(chaindata.shape[0]):
    chaindp = chaindata.iloc[i]
    chain_node_name = chaindp.nodename
    index = Nlist.index(chain_node_name)
    chainlist.append(index)
chaindata['indexer'] = range(chaindata.shape[0])
chaindata['chain_address'] = chaindata['indexer'].astype(str) + ' | '  + chaindata['chain'] + ' | ' + chaindata['address'] + ' | ' + chaindata['zipcode'].astype(str)
chainnames = np.array(chaindata['chain'].values)
chainaddress = np.array(chaindata['chain_address'].values)
chainlist = np.array(chainlist)
chaindata['node_index'] = chainlist

### Get Lengths

def computeLength(path,G):
    
    #Compute Length
    lengths = []
    for i in range(len(path)-1):
        orig = path[i]
        dest = path[i+1]
        weight = G[orig][dest]['weight']
        lengths.append(weight)

    #Return
    return lengths

### Pair Exists in data

def pairExists(source,target,chain,P):

    #Top Level miss
    exists = False
    if (source not in P):
        exists = False


    #Source in top level P
    if (source in P):
        
        #Get Target Items
        target_items = P[source]
        
        #Loop through target items to see if they are keyed on target
        for target_item in target_items:
            key = list(target_item.keys())[0]

            #If target is in P[source] determine whether chain occurs in target item
            if (target == key):
                chains,distance = target_item[key]
                if (chain in chains):
                    exists = True
    
    #Return
    return exists

### Queue Based distance computation

# #Init
# P = {}
# D = np.zeros((len(tractlist),len(chainlist)))
# totcount = 0
# runcount = 0
# totcounts = []
# runcounts = []
# durations = []
# foundpairs = []
# badpairs = []

# #Loop through cude
# for i in range(50):#$tractlist))):
#     starttime = time.time()
#     for j in range(50):#len(chainlist)):

#         #Get Source and Target
#         source = tractlist[i]
#         target = chainlist[j]
#         pair = {source,target}
#         totcount = totcount + 1

#         #Perform only if pair has not already been found
#         if (i,j) not in foundpairs:

#             #Check for exsistence of pair
#             exists = pairExists(source,target,chainnames[j],P)

#             #If pair does not exsist
#             if (exists == False):

#                 #Initial Path
#                 runcount = runcount + 1
#                 try:

#                     #Find Path
#                     paths = nx.shortest_path(G, target=target,source=source)
#                     pathlengths = computeLength(paths,G)

#                 except:
#                     paths = []
#                     pathlenghts = []
#                     badpairs.append((source,target))

#                 #Create Queue
#                 pairs = []
#                 for p1 in range(len(paths)):
#                     for p2 in range(p1+1,len(paths)):
#                         start_index = paths[p1]
#                         end_index = paths[p2]
#                         if (start_index in tractlist and end_index in chainlist):
#                             sub_length = np.sum(pathlengths[p1:p2])
#                             clist = np.where(chainlist == end_index)[0]
#                             cnames = chainnames[clist]
#                             cnames = chainaddress[clist]
#                             for c,cname in enumerate(cnames):
#                                 if (D[i,clist[c]] > 0):
#                                     if (i,clist[c]) not in foundpairs:
#                                         foundpairs.append((i,clist[c]))
#                                 else:
#                                     D[i,clist[c]] = sub_length
#                                     entity = {start_index:p1,end_index:{p2:cname}}
#                                     exists = pairExists(start_index,end_index,cname,P)
#                                     if (exists == False and entity not in pairs):
#                                         pairs.append({start_index:p1,end_index:{p2:cname}})

#                 #Empty Queue
#                 while (len(pairs) > 0):
#                     pairdic = pairs.pop(0)
#                     pairlist = list(pairdic.keys())
#                     start_index = pairlist[0]
#                     end_index = pairlist[1]
#                     p1 = pairdic[start_index]
#                     p2 = list(pairdic[end_index].keys())[0]
#                     chain = pairdic[end_index][p2]
#                     exists = pairExists(start_index,end_index,chain,P)

#                     #If node pair not already exists
#                     if (exists == False):

#                         #If Start Index not in P
#                         if (start_index not in P):
#                             P[start_index] = [{end_index:([chain],sub_length)}]

#                         #If Start Index in P
#                         else:

#                             #Find End Index in P[Start_index]
#                             end_index_list = P[start_index]
#                             found = False
#                             target_index = 0
#                             for t,target_item in enumerate(end_index_list):
#                                 key = list(target_item.keys())[0]
#                                 if (end_index == key):
#                                     found = True
#                                     target_index = t
#                                     break

#                             #If found, extract end_index and add chain
#                             if (found == True):
#                                 end_index_item = end_index_list[target_index]
#                                 mylist,mydistance = end_index_item[end_index]
#                                 mylist.append(chain)
#                                 end_index_item[end_index] = (mylist,mydistance)
#                                 end_index_list[target_index] = end_index_item
#                                 P[start_index] = end_index_list

#                             #Else, add a new item to P[start_index]
#                             else:
#                                 end_index_list = P[start_index]
#                                 end_index_list.append({end_index:([chain],sub_length)})
#                                 P[start_index] = end_index_list 


#     #Keep track of counts 
#     totcounts.append(totcount)
#     runcounts.append(runcount)
#     endtime = time.time()
#     print(i,endtime-starttime)
#     durations.append(endtime-starttime)

# foundtracts = []
# foundchains = []
# foundtogether = []
# for node1 in P:
#     tractindex = np.where(tractlist == node1)[0]
#     foundtracts.append(tractindex[0])
#     for node2 in P[node1]:
#         keys = list(node2.keys())
#         for key in keys:
#             mychains,distance = node2[key]
#             for mychain in mychains:
#                 elements = mychain.split(' | ')
#                 foundchains.append(int(elements[0]))
#                 foundtogether.append((tractindex[0],int(elements[0])))
# foundtracts = np.unique(foundtracts)
# foundchains = np.unique(foundchains)
# tractDone = float(len(foundtracts))/float(len(tractlist))*100.0
# chainDone = float(len(foundchains))/float(len(chainlist))*100.0
# totalDone = float(len(foundtogther))/float(len(tractlist)*len(chainlist))*100.0
# print('Results:')
# print(len(foundpairs))
# print(len(foundtogther))
# print(len(tractlist))
# print(len(chainlist))
# print(len(tractlist)*len(chainlist))
# print('Percent tract done: '  + str(np.round(tractDone,4)))
# print('Percent chain done: '  + str(np.round(chainDone,4)))
# print('Percent total done: '  + str(np.round(totalDone,4)))

### Find Bad Eggs

connected = []
unconnected = []
# tractlist = tractlist.tolist()
# chainlist = chainlist.tolist()
D = np.zeros((len(tractlist),len(chainlist)))
for i in range(len(tractlist)):
    starttime = time.time()
    for j in range(1):#$tractlist))):
        if (D[i,j] == 0):
            source =  tractlist[i]
            target = chainlist[j]
            try:
                path = nx.shortest_path(G,source=source,target=target)
                for p1 in range(len(path)):
                    for p2 in range(p1+1,len(path)):
                        source_sub = path[p1]
                        target_sub = path[p2]
                        if (source_sub in tractlist):
                            if (target_sub in chainlist):
                                source_index = tractlist.index(source_sub)
                                target_index = chainlist.index(target_sub)
                                D[source_index,target_index] = 1
            except:
                unconnected.append({source_index,target_index})
    endtime = time.time()
    print(endtime-starttime)

print(int(np.sum(D)))
print(len(tractlist)*len(chainlist))

### Remove Bad Eggs

tractlistnew = []
chainlistnew = []
for item in tractlist:
    if item not in badindices:
        tractlistnew.append(item)
for item in chainlist:
    if item not in badindices:
        chainlistnew.append(item)

### Simple Approach

#Init
P = {}
D = np.zeros((len(tractlist),len(chainlist)))
totcount = 0
runcount = 0
totcounts = []
runcounts = []
durations = []
foundpairs = []
badpairs = []

#Loop through cude
for i in range(1):#$tractlist))):
    starttime = time.time()
    source =  tractlistnew[i]
    targets = chainlistnew[0:1]
#     for j in range(50):#len(chainlist)):

#         #Get Source and Target
#         source = tractlist[i]
#         target = chainlist[j]
#         pair = {source,target}
#         totcount = totcount + 1
    
#         #Try to run
    paths = nx.shortest_path(G,source=source,target=targets)
    pathlengths = computeLength(paths,G)
        

#         except:
#             paths = []
#             pathlenghts = []
#             badpairs.append((source,target))
#             P[source][target] = np.sum(pathlenghts)
            
#                 #Create Queue
#                 pairs = []
#                 for p1 in range(len(paths)):
#                     for p2 in range(p1+1,len(paths)):
#                         start_index = paths[p1]
#                         end_index = paths[p2]
#                         if (start_index in tractlist and end_index in chainlist):
#                             sub_length = np.sum(pathlengths[p1:p2])
#                             clist = np.where(chainlist == end_index)[0]
#                             cnames = chainnames[clist]
#                             cnames = chainaddress[clist]
#                             for c,cname in enumerate(cnames):
#                                 if (D[i,clist[c]] > 0):
#                                     if (i,clist[c]) not in foundpairs:
#                                         foundpairs.append((i,clist[c]))
#                                 else:
#                                     D[i,clist[c]] = sub_length
#                                     entity = {start_index:p1,end_index:{p2:cname}}
#                                     exists = pairExists(start_index,end_index,cname,P)
#                                     if (exists == False and entity not in pairs):
#                                         pairs.append({start_index:p1,end_index:{p2:cname}})

#                 #Empty Queue
#                 while (len(pairs) > 0):
#                     pairdic = pairs.pop(0)
#                     pairlist = list(pairdic.keys())
#                     start_index = pairlist[0]
#                     end_index = pairlist[1]
#                     p1 = pairdic[start_index]
#                     p2 = list(pairdic[end_index].keys())[0]
#                     chain = pairdic[end_index][p2]
#                     exists = pairExists(start_index,end_index,chain,P)

#                     #If node pair not already exists
#                     if (exists == False):

#                         #If Start Index not in P
#                         if (start_index not in P):
#                             P[start_index] = [{end_index:([chain],sub_length)}]

#                         #If Start Index in P
#                         else:

#                             #Find End Index in P[Start_index]
#                             end_index_list = P[start_index]
#                             found = False
#                             target_index = 0
#                             for t,target_item in enumerate(end_index_list):
#                                 key = list(target_item.keys())[0]
#                                 if (end_index == key):
#                                     found = True
#                                     target_index = t
#                                     break

#                             #If found, extract end_index and add chain
#                             if (found == True):
#                                 end_index_item = end_index_list[target_index]
#                                 mylist,mydistance = end_index_item[end_index]
#                                 mylist.append(chain)
#                                 end_index_item[end_index] = (mylist,mydistance)
#                                 end_index_list[target_index] = end_index_item
#                                 P[start_index] = end_index_list

#                             #Else, add a new item to P[start_index]
#                             else:
#                                 end_index_list = P[start_index]
#                                 end_index_list.append({end_index:([chain],sub_length)})
#                                 P[start_index] = end_index_list 


#     #Keep track of counts 
#     totcounts.append(totcount)
#     runcounts.append(runcount)
    endtime = time.time()
    print(i,endtime-starttime)
    durations.append(endtime-starttime)

# foundtracts = []
# foundchains = []
# foundtogether = []
# for node1 in P:
#     tractindex = np.where(tractlist == node1)[0]
#     foundtracts.append(tractindex[0])
#     for node2 in P[node1]:
#         keys = list(node2.keys())
#         for key in keys:
#             mychains,distance = node2[key]
#             for mychain in mychains:
#                 elements = mychain.split(' | ')
#                 foundchains.append(int(elements[0]))
#                 foundtogether.append((tractindex[0],int(elements[0])))
# foundtracts = np.unique(foundtracts)
# foundchains = np.unique(foundchains)
# tractDone = float(len(foundtracts))/float(len(tractlist))*100.0
# chainDone = float(len(foundchains))/float(len(chainlist))*100.0
# totalDone = float(len(foundtogther))/float(len(tractlist)*len(chainlist))*100.0
# print('Results:')
# print(len(foundpairs))
# print(len(foundtogther))
# print(len(tractlist))
# print(len(chainlist))
# print(len(tractlist)*len(chainlist))
# print('Percent tract done: '  + str(np.round(tractDone,4)))
# print('Percent chain done: '  + str(np.round(chainDone,4)))
# print('Percent total done: '  + str(np.round(totalDone,4)))

### Algorithm Speed Up

plt.figure(figsize=(12,4))
plt.style.use('dark_background')
plt.subplot(1,2,1);
plt.plot(totcounts,'r',linewidth=2)
plt.plot(runcounts,'g',linewidth=2)
plt.subplot(1,2,2)
plt.plot(durations,'b');

### Node unpacker

def nodeUnpacker(item):
    
    #Unpack
    nodeindex = list(item.keys())[0]
    mychainlist,mydistance = item[nodeindex]

    #Return
    return nodeindex,mychainlist,mydistance


### Create Node space with distance a tract

data = []
mychain = 'McDonalds'
tractkeys = list(P.keys())
nodeindex1 = tractkeys[0]
for item in P[nodeindex1]:
    nodeindex2,mychainlist,mydistance = nodeUnpacker(item)
    node1 = N[nodeindex1]
    node2 = N[nodeindex2]
    elements1 = node1.split(' | ')
    elements2 = node2.split(' | ')
    lat1 = float(elements1[-3])
    lng1 = float(elements1[-2])
    lat2 = float(elements2[-3])
    lng2 = float(elements2[-2])
    chainindices = np.where(chainlist == nodeindex2)[0]
    for c,chain_address in enumerate(mychainlist):
        elements = chain_address.split(' | ')
        data.append((lat1,lng1,lat2,lng2,chain_address,elements[1],mydistance))
data = pd.DataFrame(data)
data.columns = ['lat1','lng1','lat2','lng2','chain+address','chain','distance']
data = data[data['chain'] == mychain]

### Normalize

distances = data['distance'].values
data['distance_norm'] = distances / np.max(distances)

### Load Outline

sf = shapefile.Reader('../data/shapefiles/nyc_outline/nyc_outline.shp')
streetsShapeRecs = sf.shapeRecords()
points = np.array(streetsShapeRecs[0].shape.points)
parts = streetsShapeRecs[0].shape.parts
parts.append(points.shape[0])

### Plot

B = data[['lat1','lng1']].values
X = data[['lat2','lng2']].values
rgb = np.zeros((X.shape[0],3))
rgb[:,0] = data['distance_norm'].values
rgb[:,1] = 0#data['distance_norm'].values**2
rgb[:,2] = 1-data['distance_norm'].values
offset = 0.02
skip = 1
plt.figure(figsize=(9,9))
for i in range(len(parts)-1):
    spos = parts[i]
    epos = parts[i+1]
    polygon = points[spos:epos,:]
    plt.plot(polygon[:,0],polygon[:,1],c='w',linewidth=1);
plt.scatter(X[:,0],X[:,1],c=rgb[:,:],s=60)
plt.scatter(B[0,0],B[0,1],c='w',s=60)
minval = np.min(np.concatenate((B[:1,:],X),axis=0),axis=0)
maxval = np.max(np.concatenate((B[:1,:],X),axis=0),axis=0)
plt.axis((minval[0]-offset,maxval[0]+offset,minval[1]-offset,maxval[1]+offset));

### Get Found Tracts

tracts = []
tractkeys = list(P.keys())
for key in tractkeys:
    index = np.where(tractlist == key)[0][0]
    tractdp = tractdata.iloc[index]
    lat = tractdp.lat
    lng = tractdp.lng
    tracts.append((lat,lng,tractdp.geoid))
tracts = pd.DataFrame(tracts)
tracts.columns = ['lat','lng','geoid']

### Get Chains

chains = []
tractkeys = list(P.keys())
indices = []
for key in tractkeys:
    items = P[key]
    for item in items:
        nodeindex,mychainlist,mydistance = nodeUnpacker(item)
        for mychain in mychainlist:
            elements = mychain.split(' | ')
            index = int(elements[0])
            indices.append(index)
            chaindp = chaindata.iloc[index]
            lat = chaindp.lat
            lng = chaindp.lng
            chains.append((lat,lng,chaindp.chain))
chains = pd.DataFrame(chains)
chains.columns = ['lat','lng','node']
chains = chains.drop_duplicates(keep='first')

### Load Tract Shapefile

sf = shapefile.Reader('../data/acs/nyc_tracts/nyc_tracts_cleaned.shp')
tractsShapeRecs = sf.shapeRecords()
plist = []
for i in range(len(tractsShapeRecs)):
    tractpoints = np.array(tractsShapeRecs[i].shape.points)
    tractparts = tractsShapeRecs[i].shape.parts
    tractparts.append(tractpoints.shape[0])
    plist.append((tractpoints,tractparts))

### Plot Tracts and Chains

X = tracts[['lng','lat']].values
C = chains[['lng','lat']].values
rgb = np.zeros((X.shape[0],3))
offset = 0.02
skip = 1
plt.figure(figsize=(9,9))
for m in range(len(plist)):
    tractpoints,tractparts = plist[m]
    for i in range(len(tractparts)-1):
        spos = tractparts[i]
        epos = tractparts[i+1]
        polygon = tractpoints[spos:epos,:]
        plt.plot(polygon[:,0],polygon[:,1],c='w',linewidth=0.5,zorder=1)
plt.scatter(X[:,0],X[:,1],c='r',s=10,zorder=2)
plt.scatter(C[:,0],C[:,1],c='g',s=10,zorder=3)
minval = np.min(X,axis=0)
maxval = np.max(X,axis=0)
plt.axis((minval[0]-offset,maxval[0]+offset,minval[1]-offset,maxval[1]+offset));

### Save P

X = {}
X['P'] = P
X['D'] = D
pickle.dump(X,open('../data/paths/nyc_tracts_nodes.p',"wb"))
